requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
openai==0.28.1
gunicorn==21.2.0
